requirements/base.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
psycopg2==2.9.6
pycryptodome==3.15.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
